GS-13598 modified demote API in AdminAPI to return a Future + changed timeout to be long instead of int

+++ xap-core/xap-datagrid/src/main/java/com/gigaspaces/internal/client/spaceproxy/SpaceProxyImpl.java
@@ -720,7 +720,7 @@ public class SpaceProxyImpl extends AbstractDirectSpaceProxy implements SameProx
     }
 
     @Override
-    public void demoteToBackup(int timeout, TimeUnit unit) throws DemoteFailedException {
+    public void demoteToBackup(long timeout, TimeUnit unit) throws DemoteFailedException {
         IRemoteSpace spaceImpl = getRemoteJSpace();
 
         spaceImpl.demoteToBackup(timeout, unit);
